Extracted decompose-test-utils module
