+++ decompose-test-utils/build.gradle.kts
@@ -1,0 +1,36 @@
+import com.arkivanov.gradle.bundle
+import com.arkivanov.gradle.dependsOn
+import com.arkivanov.gradle.setupBinaryCompatibilityValidator
+import com.arkivanov.gradle.setupMultiplatform
+import com.arkivanov.gradle.setupPublication
+import com.arkivanov.gradle.setupSourceSets
+
+plugins {
+    id("kotlin-multiplatform")
+    id("com.android.library")
+    id("kotlinx-serialization")
+    id("com.arkivanov.gradle.setup")
+}
+
+setupMultiplatform()
+
+android {
+    namespace = "com.arkivanov.decompose.testutils"
+}
+
+kotlin {
+    setupSourceSets {
+        all {
+            languageSettings {
+                optIn("com.arkivanov.decompose.InternalDecomposeApi")
+                optIn("com.arkivanov.decompose.ExperimentalDecomposeApi")
+            }
+        }
+
+        common.main.dependencies {
+            implementation(project(":decompose"))
+            implementation(deps.jetbrains.kotlinx.kotlinxSerializationCore)
+            implementation(deps.jetbrains.kotlinx.kotlinxSerializationJson)
+        }
+    }
+}
