+++ decompose-test-utils/src/commonMain/kotlin/com/arkivanov/decompose/testutils/TestUtils.kt
@@ -1,12 +1,13 @@
-package com.arkivanov.decompose
+package com.arkivanov.decompose.testutils
 
 import com.arkivanov.decompose.value.Value
 import com.arkivanov.essenty.statekeeper.StateKeeper
 import com.arkivanov.essenty.statekeeper.StateKeeperDispatcher
+import kotlin.reflect.KProperty
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.serializer
-import kotlin.reflect.KProperty
 
+@PublishedApi
 internal val json =
     Json {
         allowStructuredMapKeys = true
@@ -19,7 +20,7 @@ inline fun <reified T : Any> StateKeeper.register(key: String, noinline supplier
 inline fun <reified T : Any> StateKeeper.consume(key: String): T? =
     consume(key = key, strategy = serializer<T>())
 
-internal inline fun <reified T> T.serializeAndDeserialize(): T {
+inline fun <reified T> T.serializeAndDeserialize(): T {
     val serializer = serializer<T>()
 
     return json.decodeFromString(serializer, json.encodeToString(serializer, this))
