+++ decompose/src/androidUnitTest/kotlin/com/arkivanov/decompose/DefaultComponentContextBuilderTest.kt
@@ -1,6 +1,8 @@
 package com.arkivanov.decompose
 
 import com.arkivanov.decompose.router.TestInstance
+import com.arkivanov.decompose.testutils.consume
+import com.arkivanov.decompose.testutils.register
 import com.arkivanov.essenty.backhandler.BackCallback
 import com.arkivanov.essenty.instancekeeper.getOrCreate
 import org.junit.runner.RunWith
