+++ decompose/src/commonTest/kotlin/com/arkivanov/decompose/router/items/LazyItemsNavigationTest.kt
@@ -1,6 +1,5 @@
 package com.arkivanov.decompose.router.items
 
-import com.arkivanov.decompose.TestComponentContext
 import com.arkivanov.decompose.router.items.Items.ActiveLifecycleState
 import com.arkivanov.decompose.router.items.LazyComponentState.CREATED
 import com.arkivanov.decompose.router.items.LazyComponentState.DESTROYED
@@ -8,6 +7,7 @@ import com.arkivanov.decompose.router.items.LazyComponentState.PENDING
 import com.arkivanov.decompose.router.items.LazyComponentState.REMOVED
 import com.arkivanov.decompose.router.items.LazyComponentState.RESUMED
 import com.arkivanov.decompose.router.items.LazyComponentState.STARTED
+import com.arkivanov.decompose.testutils.TestComponentContext
 import kotlin.test.Test
 
 @Suppress("TestFunctionName")
